backend1/requirements.txt
Django>=5.2.6
djangorestframework>=3.15.2
django-cors-headers>=4.7.0
djangorestframework-simplejwt>=5.3.0
python-decouple>=3.8
gunicorn>=23.0.0
psycopg2-binary>=2.9.9
whitenoise>=6.8.2
dj-database-url>=2.3.0